Renamed ILayoutContainer to ILayoutGroup, and general cleanup
