+++ src/react-apps/applications/runtime/src/features/form/layout/actions/index.ts
@@ -1,5 +1,5 @@
 import { ActionCreatorsMapObject, bindActionCreators } from 'redux';
-import { ILayoutComponent, ILayoutContainer } from '..';
+import { ILayoutComponent, ILayoutGroup } from '..';
 import { store } from '../../../../store';
 import * as FetchForm from './fetch';
 import * as UpdateFormLayout from './update';
@@ -8,7 +8,7 @@ export interface IFormLayoutActions extends ActionCreatorsMapObject {
   fetchFormLayout: (url: string) => FetchForm.IFetchFormLayout;
   fetchFormLayoutFulfilled: (layout: any) => FetchForm.IFetchFormLayoutFulfilled;
   fetchFormLayoutRejected: (error: Error) => FetchForm.IFetchFormLayoutRejected;
-  updateFormLayout: (formLayoutElement: ILayoutComponent | ILayoutContainer, index: number)
+  updateFormLayout: (formLayoutElement: ILayoutComponent | ILayoutGroup, index: number)
     => UpdateFormLayout.IUpdateFormLayout;
 }
