+++ src/react-apps/applications/runtime/tsconfig.json
@@ -26,7 +26,7 @@
     "rootDir": "../"
   },
   "include": [
-    "./runtime/src/**/*",
+    "./src/**/*",
   ],
   "exclude": [
     "node_modules",
